♻️ refactor: 페키지 구조 수정 & mapping 테이블에 BaseEntity 상속
