+++ src/main/java/com/example/umc/study/domain/Post.java
@@ -1,6 +1,6 @@
 package com.example.umc.study.domain;
 
-import com.example.umc.study.common.BaseEntity;
+import com.example.umc.study.domain.common.BaseEntity;
 import com.example.umc.study.domain.mapping.PostCategory;
 import jakarta.persistence.*;
 import lombok.*;
